requirements.txt
Django==3.1
django-filter
djangorestframework
djangorestframework-simplejwt==4.6.0
dnspython==1.16.0
httplib2==0.15.0
requests==2.22.0
django-redis
Pillow==8.1.0
python-dotenv
psycopg2
django-cors-headers==3.6.0
